Fix and improve performance of installation of latest xdebug
The pecl resource itself doesn't run properly on ubuntu >= 12.10
in this version of the php cookbook, and updating the channel every
time when we want to install a known version was wasteful anyway.

For now, switch to running a bash script to verify the installed
pecl version and only update if it doesn't match what we require.